Rework solar system to better support distant planet sprites
Planets are now “always visible” as small specs that you might see in
the sky mistaken for a stars… until they appear to move as you travel.
The sun will be addressed in a separate addition. The planets
themselves will be disabled when far enough away and enabled before
overtaking the “spec” to help with performance.